Tried to associate room member and user.

+++ healthcare/views.py
@@ -1,7 +1,7 @@
 from django.shortcuts import render
 from rest_framework.decorators import api_view
-from .models import Doctor, Patient
-from .serializers import DoctorSerializer, PatientSerializer
+from .models import Doctor, Patient, NewUser
+from .serializers import DoctorSerializer, PatientSerializer, NewUserSerializer
 from rest_framework.response import Response
 from django.http import HttpResponse
 # Create your views here.
@@ -85,6 +85,48 @@ def doctors(request):
         data = request.data
         try:
             obj = Doctor.objects.get(id=data['id'])
+            name = obj.username
+            obj.delete()
+            return Response({'message' : f"{name} deleted successfully."})
+        except Exception as e:
+            return Response({'message' : f"Error is {e}"})
+        
+@api_view(['GET', 'POST', 'PUT', 'PATCH', 'DELETE'])
+def new_Users(request):
+    if request.method == 'GET':
+        patients = NewUser.objects.all()
+        serializer = NewUserSerializer(patients, many=True)
+        return Response(serializer.data)
+    
+    elif request.method == 'POST':
+        data = request.data
+        serializer = NewUserSerializer(data=data)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data)
+        return Response(serializer.errors)
+    
+    elif request.method == 'PUT':
+        data = request.data
+        serializer = NewUserSerializer(data=data)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data)
+        return Response(serializer.errors)
+    
+    elif request.method == 'PATCH':
+        data = request.data
+        obj = NewUser.objects.get(id=data['id'])
+        serializer = NewUserSerializer(obj, data=data, partial=True)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data)
+        return Response(serializer.errors)
+    
+    elif request.method == 'DELETE':
+        data = request.data
+        try:
+            obj = NewUser.objects.get(id=data['id'])
             name = obj.username
             obj.delete()
             return Response({'message' : f"{name} deleted successfully."})
